fix(claude-code): catch up to v2.1.117 + tighten clippy gate (#756)
* fix(claude-code): catch up to v2.1.117 (Monitor tool, xhigh effort, color field)

Closes #745.

Triage of upstream Claude Code releases v2.1.91 through v2.1.117 (27-release
window, 20 with notes) surfaced three structural changes that were causing
agnix to false-positive on valid Claude Code configs:

- Monitor tool (added v2.1.98) is now in KNOWN_AGENT_TOOLS, so an agent with
  tools: [Monitor] no longer triggers CC-AG-009/CC-AG-010.
- xhigh effort level (added v2.1.111 for Opus 4.7) is now in
  VALID_EFFORT_VALUES (agent) and VALID_EFFORT_LEVELS (skill), so
  effort: xhigh no longer triggers CC-AG-014/CC-SK-018.
- color field (display color, predates this window but exposed during
  upstream cross-check) is now a typed field on AgentSchema and listed in
  KNOWN_AGENT_FIELDS, so color: blue no longer triggers CC-AG-019.

Documentation: bumped verified_on for affected rules to 2026-04-22; added a
note on CC-AG-011 and CC-AG-019 that hooks/mcpServers in agent frontmatter
also apply to --agent main-thread sessions (v2.1.116/v2.1.117); bumped
RESEARCH-TRACKING.md Last Reviewed for Claude Code.

Tests added: test_monitor_tool_accepted, test_color_field_accepted, and
extended test_effort_valid_values / test_cc_sk_018_valid_effort_values to
cover xhigh.

Verified: cargo test --workspace passes (4500+ tests, 0 failures);
cargo clippy --workspace --all-features -- -D warnings clean.

Out of scope (will land in a follow-up PR before #747): output-style
validator (new file type, .claude/output-styles/*.md with the
keep-coding-instructions field added in v2.1.94).

* chore(ci): tighten clippy gate to --all-targets and fix latent warnings

The CI clippy step and local pre-push hook ran `cargo clippy --workspace
--all-features` without `--all-targets`, so warnings in test, bench, and
example code never failed the gate. Adding `--all-targets` exposed eight
latent warnings, all fixed in this commit:

- agnix-cli/src/main.rs: items_after_test_module - moved
  `mod resolve_fix_mode_tests` to end of file
- agnix-cli/src/sarif.rs: unnecessary_literal_unwrap on intentional fallback
  test - marked `#[allow]` since the test exercises that exact path
- agnix-cli/tests/kiro_fixture_inventory.rs: needless `<'a>` lifetime - dropped
- agnix-core/benches/validation.rs: deprecated `criterion::black_box` ->
  `std::hint::black_box`
- agnix-core/src/rules/cross_platform.rs: needless `&content` (already &str)
- agnix-core/src/rules/project_level.rs: needless `&` in two `.contains()`
  calls
- agnix-core/tests/api_contract.rs: `Ok(42).unwrap()` on intentional contract
  test - marked `#[allow]`
- agnix-lsp/src/backend/tests.rs: unnecessary `(i + 2) as i32` cast
- agnix-lsp/tests/lsp_integration.rs: unused `DiagnosticLevel` and
  `LintConfig` imports - removed

Verified: `cargo clippy --workspace --all-targets --all-features -- -D warnings`
passes with these changes.

* fix(clippy): sort_by_key with Reverse for the remaining sort_by sites

Rust 1.95 / clippy 1.95 in CI flagged 11 more  errors
(pattern `refs.sort_by(|a, b| b.start_byte.cmp(&a.start_byte))` -> use
`sort_by_key(|b| std::cmp::Reverse(b.start_byte))`) that the previous
clippy bundle commit did not catch because the local toolchain was 1.92.

10 sites in agnix-core/src/fixes.rs and 1 in agnix-core/src/rules/xml.rs.
Pure mechanical refactor, no behavior change. Verified: cargo clippy
--workspace --all-targets --all-features -- -D warnings now passes;
3409 lib tests pass.

* fix: address review comments (round 1) + cfg(unix) on lsp test imports

Reviewer feedback (PR #756 round 1):

- Gemini #2: dedupe agent VALID_EFFORT_VALUES with skill VALID_EFFORT_LEVELS;
  agent.rs now imports from schemas::skill so the two stay in lock-step.
- Gemini #3: extend test_monitor_tool_accepted to also cover CC-AG-010
  (disallowedTools list) since both validators share KNOWN_AGENT_TOOLS.
- Gemini #4: drop redundant test_effort_valid_values; the pre-existing
  test_cc_ag_014_all_valid_effort_values already iterates VALID_EFFORT_LEVELS.
- Gemini #5: skill tests.rs now iterates VALID_EFFORT_LEVELS instead of
  hardcoding the list.
- Copilot #6: drop redundant #[allow(unnecessary_literal_unwrap)] on the
  Ok(42) declaration line in api_contract.rs (the actual unwrap is on the
  next line and already has its own allow).
- Copilot #7: move the v2.1.116/v2.1.117 --agent-main-thread scope note out
  of CC-AG-011 and CC-AG-019 (which check structure, not loading semantics);
  consolidate as a single scope note at the top of the SUBAGENTS section.

CI fix (Linux/macOS only):
- agnix-lsp/tests/lsp_integration.rs: gate `use agnix_core::{DiagnosticLevel,
  LintConfig}` with #[cfg(unix)] so it follows the unix-gated test that uses
  it. The import was removed by the clippy bundle (Windows-local clippy
  thought it was unused since the test doesn't compile on Windows), but
  Linux/macOS CI does compile the test and needed the import.

Verified locally: cargo clippy --workspace --all-targets --all-features
-- -D warnings clean; 3408 lib tests pass; rules.json parity test passes.

* docs(changelog): reference VALID_EFFORT_LEVELS not the removed VALID_EFFORT_VALUES

Per Copilot review (PR #756 round 2): the CHANGELOG entry referenced
`VALID_EFFORT_VALUES` and CC-AG-011/CC-AG-019 notes that no longer exist
after the round-1 dedup and the CC-AG-019 note rephrasing. Update to
match the actual implementation: shared `schemas::skill::VALID_EFFORT_LEVELS`
constant + scope note at the SUBAGENTS section header.

+++ crates/agnix-lsp/src/backend/tests.rs
@@ -3276,7 +3276,7 @@ async fn test_stress_concurrent_hover_during_validation() {
                     .did_change(DidChangeTextDocumentParams {
                         text_document: VersionedTextDocumentIdentifier {
                             uri,
-                            version: (i + 2) as i32,
+                            version: i + 2,
                         },
                         content_changes: vec![TextDocumentContentChangeEvent {
                             range: None,
